refactor: integrating chatPannelHeader into chatPannel
