+++ apps/masterbots.ai/components/routes/chat/chat-panel.tsx
@@ -1,31 +1,36 @@
-//* ChatPanel is the main chat UI component, managing user inputs and chat actions like reloading, stopping, and sending prompts.
-
-import { ChatPanelHeader } from '@/components/routes/chat/chat-panel-header'
-import { PromptForm } from '@/components/routes/chat/prompt-form'
-import { useThread } from '@/lib/hooks/use-thread'
-import { cn } from '@/lib/utils'
+import * as React from 'react'
 import type { UseChatHelpers } from 'ai/react'
 import type { Chatbot } from 'mb-genql'
+import { GlobeIcon } from 'lucide-react'
+import { appConfig } from 'mb-env'
+
+import { PromptForm } from '@/components/routes/chat/prompt-form'
+import { ChatShareDialog } from '@/components/routes/chat/chat-share-dialog'
+import { ButtonScrollToBottom } from '@/components/shared/button-scroll-to-bottom'
+import { FontSizeSelector } from '@/components/shared/font-size-selector'
+import { Button } from '@/components/ui/button'
+import { Checkbox } from '@/components/ui/checkbox'
+import { IconRefresh, IconShare, IconStop } from '@/components/ui/icons'
+import { Label } from '@/components/ui/label'
+import { Separator } from '@/components/ui/separator'
+import { Switch } from '@/components/ui/switch'
+import { usePowerUp } from '@/lib/hooks/use-power-up'
+import { useThread } from '@/lib/hooks/use-thread'
+import { cn } from '@/lib/utils'
 
 export interface ChatPanelProps
   extends Pick<
     UseChatHelpers,
-    | "append"
-    | "isLoading"
-    | "reload"
-    | "messages"
-    | "stop"
-    | "input"
-    | "setInput"
+    'append' | 'isLoading' | 'reload' | 'messages' | 'stop' | 'input' | 'setInput'
   > {
-  scrollToBottom: () => void // Function to scroll chat to the bottom
-  id?: string // Chat ID, used in message operations
-  title?: string // Chat title, displayed in the header
-  chatbot?: Chatbot // Chatbot configuration for enabling/disabling prompt form
-  showReload?: boolean // Displays reload button when true
-  placeholder: string // Placeholder text for the input field
-  isAtBottom?: boolean // Indicates if the chat is scrolled to the bottom
-  className?: string // Optional custom class for styling the panel
+  scrollToBottom: () => void
+  id?: string
+  title?: string
+  chatbot?: Chatbot
+  showReload?: boolean
+  placeholder: string
+  isAtBottom?: boolean
+  className?: string
 }
 
 export function ChatPanel({
@@ -45,7 +50,10 @@ export function ChatPanel({
   scrollToBottom,
   className,
 }: ChatPanelProps) {
-  const { isOpenPopup, loadingState } = useThread() // State to control popup visibility
+  const { isOpenPopup, loadingState, webSearch, setWebSearch } = useThread()
+  const { isPowerUp, togglePowerUp } = usePowerUp()
+  const [shareDialogOpen, setShareDialogOpen] = React.useState(false)
+  const webSearchRef = React.useRef(null)
 
   return (
     <div
@@ -55,21 +63,122 @@ export function ChatPanel({
         'bg-gradient-to-b from-background/50 to-background',
         'dark:from-background/0 dark:to-background/80',
         'lg:pl-[250px] xl:pl-[300px]',
-        className
+        className,
       )}
     >
       <div className="relative w-full mx-auto">
-        <ChatPanelHeader
-          id={id}
-          title={title}
-          isLoading={isLoading}
-          stop={stop}
-          reload={reload}
-          messages={messages}
-          showReload={showReload}
-          scrollToBottom={scrollToBottom}
-          isAtBottom={isAtBottom}
-        />
+        {/* Header Section */}
+        <div className="flex flex-col items-center justify-between w-full px-2 py-3.5 space-y-2 bg-background md:flex-row md:space-y-0">
+          <div className="flex items-center justify-between w-full gap-4 mx-2">
+            <div className="flex items-center space-x-4">
+              {/* Power-Up Switch */}
+              <div className="flex items-center space-x-2">
+                <Switch id="power-up" checked={isPowerUp} onCheckedChange={togglePowerUp} />
+                <Label htmlFor="power-up" className="text-sm font-normal">
+                  Power-Up
+                </Label>
+              </div>
+
+              {appConfig.features.webSearch && (
+                <>
+                  <Separator orientation="vertical" className="h-6" />
+                  {/* Web Search Checkbox */}
+                  <Checkbox
+                    custom
+                    name="webSearch"
+                    id="webSearch"
+                    value={webSearch ? 'checked' : 'unchecked'}
+                    ref={webSearchRef}
+                    onClick={(value) => setWebSearch(!value)}
+                    className="transition-all delay-100 h-auto w-auto inline-flex items-center gap-2 border-muted p-0.5 data-[state=checked]:border-accent/50 data-[state=checked]:bg-accent/25 rounded-full"
+                    checkboxconfig={{
+                      check: (
+                        <>
+                          <div className="bg-accent rounded-full -m-[2px] mr-1 p-0.5">
+                            <GlobeIcon className="size-7 text-accent-foreground" />
+                          </div>
+                          <Label
+                            htmlFor="webSearch"
+                            className="mr-2 text-xs leading-none text-nowrap"
+                          >
+                            Web search enabled
+                          </Label>
+                        </>
+                      ),
+                      uncheck: (
+                        <>
+                          <GlobeIcon className="opacity-75 size-7" />
+                          <span className="sr-only">Web search disabled</span>
+                        </>
+                      ),
+                    }}
+                  />
+                </>
+              )}
+
+              <Separator orientation="vertical" className="h-6" />
+
+              {/* Font Size Selector */}
+              <div className="flex items-center">
+                <FontSizeSelector />
+              </div>
+            </div>
+
+            {/* Right side controls */}
+            <div className="flex items-center gap-2">
+              {showReload &&
+              (isLoading ||
+                loadingState?.match(/^(processing|digesting|generating|polishing|ready)/)) ? (
+                <>
+                  {loadingState && (
+                    <div className="flex items-center justify-between gap-4">
+                      <div className="flex items-center space-x-1 drop-shadow-lg">
+                        <div className="rounded-full size-2 bg-primary animate-pulse" />
+                        <div className="rounded-full size-2 bg-primary animate-pulse" />
+                        <div className="rounded-full size-2 bg-primary animate-pulse" />
+                      </div>
+                      <p className="text-sm font-bold text-primary drop-shadow-lg">
+                        {loadingState}
+                      </p>
+                    </div>
+                  )}
+                  <Button variant="outline" onClick={stop} className="bg-background">
+                    <IconStop className="mr-2" />
+                    Stop generating
+                  </Button>
+                </>
+              ) : (
+                messages?.length >= 2 && (
+                  <>
+                    <Button variant="outline" className="relative group" onClick={() => reload()}>
+                      <IconRefresh className="mr-2 transition-all" />
+                      Regenerate response
+                    </Button>
+                    {id && title && (
+                      <>
+                        <Button variant="outline" onClick={() => setShareDialogOpen(true)}>
+                          <IconShare className="mr-2" />
+                          Share
+                        </Button>
+                        <ChatShareDialog
+                          onCopy={() => setShareDialogOpen(false)}
+                          chat={{
+                            id,
+                            title,
+                            messages,
+                          }}
+                        />
+                      </>
+                    )}
+                  </>
+                )
+              )}
+            </div>
+          </div>
+          <ButtonScrollToBottom scrollToBottom={scrollToBottom} isAtBottom={isAtBottom} />
+        </div>
+
+        {/* Prompt Form Section */}
         <div
           className={cn(
             'relative flex flex-col w-full',
@@ -77,7 +186,7 @@ export function ChatPanel({
             'border-t shadow-lg bg-background',
             'dark:border-zinc-800 border-zinc-200',
             isOpenPopup ? 'dark:border-mirage border-iron' : '',
-            'min-h-[64px] sm:min-h-[80px]'
+            'min-h-[64px] sm:min-h-[80px]',
           )}
         >
           <PromptForm
@@ -85,8 +194,8 @@ export function ChatPanel({
               await append({
                 id,
                 content: value,
-                role: "user",
-              });
+                role: 'user',
+              })
             }}
             // biome-ignore lint/complexity/noExtraBooleanCast: <explanation>
             disabled={!Boolean(chatbot) || isLoading}
